% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_genogeo.R
\name{app_genogeo}
\alias{app_genogeo}
\title{Shiny application for GenoGeoGrapher}
\usage{
app_genogeo(db_list = NULL, reporting_panel = TRUE)
}
\arguments{
\item{db_list}{A named list of databases of reference populations. 
Each component is expected to be returned from \code{pops_to_DB}.}

\item{reporting_panel}{Logical. Should report generate and download be available after sample analysis.}
}
\description{
Shiny application for GenoGeoGrapher
}
