% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plots.R
\name{bar_colour}
\alias{bar_colour}
\title{bar_colour}
\usage{
bar_colour(df, alpha = 1)
}
\arguments{
\item{df}{A data.frame with at least three coloums. The first column is the logP, the second logical (z_score accept/reject), the third a unique naming column.}

\item{alpha}{Should the alpha opacity be applied? And what value, 1 = solid, 0 = transparent.}
}
\description{
Creates the colour scale for the accepted and rejected populations based on z-score and the log likelihood (log P).
}
