% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_tilt.R
\name{exponent_tilt}
\alias{exponent_tilt}
\title{P-values from Importing Sampling using Exponential tilting}
\usage{
exponent_tilt(x0, x1, n, p_limit = 0.1, B = 500, return_all = FALSE)
}
\arguments{
\item{x0}{Allele count of profile}

\item{x1}{Population allele count}

\item{n}{Sampled alleles in total in population}

\item{p_limit}{Upper limit to which we use the normal approximation}

\item{B}{An integer specifying the number of importance samples.}

\item{return_all}{Default is FALSE. If TRUE: Returns p-value, standard deviation, and method (and diagnostics).}
}
\value{
If \code{return_all= FALSE} the p-value is returned. Otherwise list of elements (see \code{return_all}) is returned.
}
\description{
P-values from Importing Sampling using Exponential tilting
}
\details{
The method of importance sampling described in Tvedebrink et al (2018), Section 2.3 is implemented.
It relies on exponential tilting of the proposal distribution using in the importance sampling.
}
