% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genogeographer.R
\docType{package}
\name{genogeographer}
\alias{genogeographer}
\alias{genogeographer-package}
\title{genogeographer: Methods for analysing forensic Ancestry Informative Markers}
\description{
The genogeographer package provides:
genogeo()
}
\section{genogeo functions}{

See ?genogeo
}

