% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kidd_loci}
\alias{kidd_loci}
\title{Kenn Kidd Lab markers}
\format{List of 55 markers
\describe{
  \item{locus}{Locus/Marker names}
}}
\source{
K.K. Kidd et al.Progress toward an efficient panel of SNPs for ancestry inference.
Forensic Science International: Genetics 10 (2014) 23–32
}
\usage{
kidd_loci
}
\description{
List of markers identified by Kenn Kidd lab.
}
\keyword{datasets}
