% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{main_alleles}
\alias{main_alleles}
\title{AIMs markers in Precision ID Ancestry Panel (Thermo Fisher Scientific)}
\format{List of 164 markers
\describe{
  \item{locus}{Locus/Marker names}
  \item{main_allele}{The main allele (alleles are in lexicographic order)}
  \item{other_allele}{The other variant}
}}
\usage{
main_alleles
}
\description{
List of markers with their main and alternative allele. 
The markers is the union of Seldin's and Kidd's markers.
}
\keyword{datasets}
