% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_AA_x0.R
\name{profile_AA_x0}
\alias{profile_AA_x0}
\title{Function that compute the genotype probability for each population (rows in df)}
\usage{
profile_AA_x0(AA_profile, df, select = c("locus", "x0"),
  keep_dropped = FALSE)
}
\arguments{
\item{AA_profile}{A tibble/data.frame with columns 'locus', 'A1' and 'A2' holding the separated version of a genotype, eg. AG -> A1: A, A2: G}

\item{df}{The database with main alleles per locus}

\item{select}{Which columns to return}

\item{keep_dropped}{Logical. Keep the non-matching alleles (compared to `db`) and those with genotype `NN`}
}
\description{
Function that compute the genotype probability for each population (rows in df)
}
