% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_AIMs_profile.R
\name{random_AIMs_profile}
\alias{random_AIMs_profile}
\title{Simulate a random AIMs profile}
\usage{
random_AIMs_profile(df, grouping = "pop", population = NULL,
  n = FALSE, keep_pop = FALSE)
}
\arguments{
\item{df}{Database of reference profiles as returned by \code{pops_to_DB}}

\item{grouping}{Simualte from \code{pop} (default) or \code{meta}.}

\item{population}{The population to sample from. If NULL chosen at random.}

\item{n}{Use numbers of samples as weights to choose the population randomly}

\item{keep_pop}{Keep information on population}
}
\description{
Use the information from \code{pops_to_DB} to simulate a profile from a random or given population.
The sampling is done with respect to the null hypothesis, such that the total count is adjusted accordingly.
For further details see Tvedebrink et al (2018), Section 3.1 (Simulations).
}
\author{
Torben Tvedebrink \email{tvede@math.aau.dk}
}
