% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seldin_loci}
\alias{seldin_loci}
\title{Seldin Lab markers}
\format{List of 122 markers
\describe{
  \item{locus}{Locus/Marker names}
}}
\source{
Kosoy et al.  
Ancestry Informative Marker Sets for Determining Continental Origin and Admixture Proportions in Common Populations in America. 
HUMAN MUTATION, Vol. 30, No. 1, 69–78, 2009.
}
\usage{
seldin_loci
}
\description{
List of markers identified by Seldin lab.
}
\keyword{datasets}
