% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_pops.R
\name{simulate_pops}
\alias{simulate_pops}
\title{Simulate random populations}
\usage{
simulate_pops(pop_n = 100, pop_names = NULL, pop_totals = NULL,
  aims_n = 50, aims_names = NULL)
}
\arguments{
\item{pop_n}{Number of populations to simulate}

\item{pop_names}{Their names. If NULL: The names are "pop_001" through "pop_pop_n"}

\item{pop_totals}{How many observations/sampled individuals per population. If one number this is used as parameter in a Poisson distribution}

\item{aims_n}{Number of AIMs}

\item{aims_names}{Their names. If NULL: The names are "rs_001" through "rs_aims_n"}
}
\description{
Simulate random populations
}
\author{
Torben Tvedebrink \email{tvede@math.aau.dk}
}
